+++ source/ArrowKeyStepper/ArrowKeyStepper.js
@@ -4,7 +4,6 @@ import type {RenderedSection} from '../Grid';
 import type {ScrollIndices} from './types';
 
 import * as React from 'react';
-import {polyfill} from 'react-lifecycles-compat';
 
 /**
  * This HOC decorates a virtualized component and responds to arrow-key events by scrolling one row or column at a time.
@@ -189,6 +188,5 @@ class ArrowKeyStepper extends React.PureComponent<Props, State> {
   }
 }
 
-polyfill(ArrowKeyStepper);
 
 export default ArrowKeyStepper;