+++ source/Collection/CollectionView.js
@@ -2,7 +2,6 @@
 import clsx from 'clsx';
 import PropTypes from 'prop-types';
 import * as React from 'react';
-import {polyfill} from 'react-lifecycles-compat';
 import createCallbackMemoizer from '../utils/createCallbackMemoizer';
 import getScrollbarSize from 'dom-helpers/scrollbarSize';
 
@@ -591,6 +590,5 @@ class CollectionView extends React.PureComponent {
   };
 }
 
-polyfill(CollectionView);
 
 export default CollectionView;